+++ src/uncrustify_types.h
@@ -280,19 +280,19 @@ struct chunk_t
    indent_ptr_t indent;
    c_token_t    type;
    c_token_t    parent_type;      /* usually CT_NONE */
-   UINT32       orig_line;
-   UINT32       orig_col;
-   UINT32       orig_col_end;
+   size_t       orig_line;
+   size_t       orig_col;         /* is always > 0 */
+   UINT32       orig_col_end;     /* is always > 1 */
    UINT32       orig_prev_sp;     /* whitespace before this token */
    UINT64       flags;            /* see PCF_xxx */
    int          column;           /* column of chunk */
-   int          column_indent;    /* if 1st on a line, set to the 'indent'
+   size_t       column_indent;    /* if 1st on a line, set to the 'indent'
                                    * column, which may be less than the real column
                                    * used to indent with tabs */
-   int          nl_count;         /* number of newlines in CT_NEWLINE */
-   int          level;            /* nest level in {, (, or [ */
-   int          brace_level;      /* nest level in braces only */
-   int          pp_level;         /* nest level in #if stuff */
+   size_t       nl_count;         /* number of newlines in CT_NEWLINE */
+   size_t       level;            /* nest level in {, (, or [ */
+   size_t       brace_level;      /* nest level in braces only */
+   size_t       pp_level;         /* nest level in #if stuff */
    bool         after_tab;        /* whether this token was after a tab */
    unc_text     str;              /* the token text */
 };
@@ -357,9 +357,9 @@ struct lookup_entry_t
 
 struct align_t
 {
-   int       col;
+   size_t    col;
    c_token_t type;
-   int       len;    // of the token + space
+   size_t    len;    // of the token + space
 };
 
 struct file_mem
@@ -442,7 +442,7 @@ struct cp_data
    int                pass_count;
 
    struct align_t     al[80];
-   int                al_cnt;
+   size_t             al_cnt;
    bool               al_c99_array;
 
    bool               warned_unable_string_replace_tab_chars;
